alert skills

+++ src/app/components/skills/skills.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { faPenToSquare, faTrashCan ,faPlus} from '@fortawesome/free-solid-svg-icons';
 import { Skill } from 'src/app/model/skill';
+import Swal from 'sweetalert2';
 import { SkillService } from '../../services/skill.service';
 import { TokenService } from '../../services/token.service';
 @Component({
@@ -44,11 +45,46 @@ export class SkillsComponent implements OnInit {
 
   delete( id: number ) {
     if( id != undefined ){
-      this.skillService.delete(id).subscribe( data => {
-        this.loadSkills();
-      }, err => {
-        alert("no se pudo borrar la skill")
-      });
+
+      Swal.fire({
+        title: 'Estas seguro?',
+        text: "Mo podrás revertir esto",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si,Borrar!',
+        cancelButtonText: 'Cancelar'
+      }).then((result) => {
+
+        if (result.isConfirmed) {
+          Swal.fire({
+            
+            title: 'Borrando skill',
+            text: 'Espere...',
+            showConfirmButton: false
+          })
+          this.skillService.delete(id).subscribe( data => {
+            Swal.fire({
+              icon: 'success',
+              title: 'Skill Borrado',
+              timer: 1500,
+              showConfirmButton: false
+            })
+            this.loadSkills();
+          }, err => {
+            Swal.fire({
+
+              icon: 'error',
+              title: 'no se pudo borrar la Skill',
+              showConfirmButton: false,
+              timer: 2500
+            })
+          });
+
+        }
+      })
+
     }
   }
 }
